# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": []
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMS Sprint 0 API",
    "description": "Frozen Sprint 0 operational API contract. Tenant context is derived from the bearer token.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "tenantCode": {
            "type": "string",
            "example": "ACME"
          },
          "loginIdentity": {
            "type": "string",
            "example": "owner@example.com"
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "minLength": 8,
            "maxLength": 256
          }
        },
        "required": [
          "tenantCode",
          "loginIdentity",
          "password"
        ]
      },
...
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "APPROVED_RECONCILIATION_NOT_FOUND",
              "AUTHENTICATED_USER_INACTIVE",
              "AUTHENTICATION_REQUIRED",
              "CASH_COUNT_STAFF_MISMATCH",
              "CASH_HANDOVER_NOT_FOUND",
              "CASH_VARIANCE_REASON_REQUIRED",
              "CONTAINER_DUE_AUTHORIZATION_REQUIRED",
              "CRITICAL_RECONCILIATION_EXCEPTION",
              "CUSTOMER_NOT_ACTIVE",
              "CUSTOMER_NOT_FOUND",
              "CUSTOMER_NOT_PENDING_APPROVAL",
              "CUSTOMER_PRICE_PERIOD_OVERLAP",
              "DAMAGE_QUANTITY_REQUIRED",
              "DAMAGE_RATE_NOT_CONFIGURED",
              "DAMAGE_RATE_PERIOD_OVERLAP",
              "DAMAGE_RESOLUTION_REQUIRED",
              "DUPLICATE_EVENT_PRODUCT",
              "EMPTY_EVENT_PRODUCT",
              "EMPTY_RETURN_NOT_ALLOWED",
              "EXCESS_EMPTY_RESOLUTION_REQUIRED",
              "EXCHANGE_RATIO_NOT_ALLOWED",
              "EXCHANGE_RATIO_REQUIRED",
              "FIXED_PLANNED_DEFAULT_QTY_REQUIRED",
              "FIXED_PLANNED_PLANNED_QTY_REQUIRED",
              "FOLLOW_UP_NOT_ALLOWED",
              "FOLLOW_UP_QUANTITY_MISMATCH",
              "FOLLOW_UP_REASON_REQUIRED",
              "HANDOVER_EXCEEDS_STAFF_CASH",
              "HANDOVER_RECIPIENT_NOT_FOUND",
              "HTTP_ERROR",
              "INSUFFICIENT_VEHICLE_STOCK",
              "INTERNAL_ERROR",
              "INVALID_ACCESS_TOKEN",
              "INVALID_CASH_HANDOVER_STATUS",
              "INVALID_CREDENTIALS",
              "INVALID_DEFAULT_QTY",
              "INVALID_EFFECTIVE_PERIOD",
              "INVALID_EXCHANGE_RATIO",
              "INVALID_FORECAST_QTY",
              "INVALID_PLANNED_QTY",
              "INVALID_POST_RECONCILIATION_ADJUSTMENT",
              "INVALID_RECONCILIATION_LOCATION",
              "INVALID_RECONCILIATION_STATUS",
              "INVALID_REFERENCE",
              "INVALID_ROUTE_CUSTOMER",
              "INVALID_ROUTE_STOP_PRODUCT",
              "INVALID_STOP_SOURCE",
              "INVALID_TRIP_STAFF_PERIOD",
              "INVALID_TRIP_STATUS",
              "INVENTORY_STATE_CONFIGURATION_MISSING",
              "LATE_EVENT_PAYMENT_NOT_SUPPORTED",
              "NEGATIVE_DERIVED_CASH",
              "NEGATIVE_DERIVED_STOCK",
              "NOT_FOUND",
              "OPEN_RECONCILIATION_EXCEPTION",
              "PARTIAL_STOP_PRODUCT_NOT_FOUND",
              "PARTY_PRODUCT_NOT_CONFIGURED",
              "PARTY_PRODUCT_NOT_FOUND",
              "PENDING_POST_RECONCILIATION_ADJUSTMENT",
              "POST_RECONCILIATION_ADJUSTMENT_NOT_FOUND",
              "PRICE_NOT_CONFIGURED",
              "PRODUCT_NOT_ACTIVE",
              "PRODUCT_NOT_FOUND",
              "PRODUCT_PRICE_PERIOD_OVERLAP",
              "QUANTITY_MODE_MISMATCH",
              "RECONCILIATION_ALREADY_EXISTS",
              "RECONCILIATION_COUNTS_REQUIRED",
              "RECONCILIATION_EXCEPTION_NOT_FOUND",
              "RECONCILIATION_EXCEPTION_NOT_OPEN",
              "RECONCILIATION_NOT_APPROVED",
              "RECONCILIATION_NOT_EDITABLE",
              "RECONCILIATION_NOT_FOUND",
              "RESOURCE_CONFLICT",
              "ROLE_NOT_AUTHORIZED",
              "ROLE_NOT_FOUND",
              "ROUTE_NOT_FOUND",
              "ROUTE_STOP_NOT_FOUND",
              "ROUTE_STOP_ORDER_CONFLICT",
              "ROUTE_STOP_REORDER_NOT_ALLOWED",
              "SELF_APPROVAL_NOT_ALLOWED",
              "SELF_VARIANCE_APPROVAL_NOT_ALLOWED",
              "STAFF_LOGIN_REQUIRED",
              "STAFF_NOT_ACTIVE",
              "STAFF_NOT_ASSIGNED_TO_TRIP",
              "STAFF_NOT_FOUND",
              "STOCK_DIMENSION_NOT_FOUND",
              "STOCK_VARIANCE_REASON_REQUIRED",
              "STOP_ALREADY_HANDLED",
              "STOP_NOT_FOUND",
              "STOP_REORDER_NOT_ALLOWED",
              "TRIP_NOT_COMPLETED",
              "TRIP_NOT_FOUND",
              "TRIP_STAFF_NOT_FOUND",
              "TRIP_STOP_NOT_FOUND",
              "TRIP_STOP_ORDER_CONFLICT",
              "TRIP_STOP_PRODUCT_NOT_FOUND",
              "TRIP_VEHICLE_REQUIRED",
              "UNRESOLVED_TRIP_STOPS",
              "USER_IDENTITY_REQUIRED",
              "USER_NOT_FOUND",
              "VALIDATION_FAILED",
              "VEHICLE_INVENTORY_LOCATION_REQUIRED",
              "VEHICLE_NOT_AVAILABLE",
              "VEHICLE_NOT_FOUND"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AuthRoleData": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "additionalProperties": true
      },
      "AuthStaffData": {
        "type": "object",
        "properties": {
          "staffId": {
            "type": "string",
            "example": "1"
          },
          "employeeCode": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "staffType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "staffId",
          "employeeCode",
          "name",
          "staffType",
          "status"
        ],
        "additionalProperties": true
      },
      "CurrentUserData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "example": "1"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthRoleData"
            }
          },
          "staff": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthStaffData"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "displayName",
          "tenantId",
          "roles",
          "staff"
        ],
        "additionalProperties": true
      },
      "LoginData": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "integer",
            "example": 1800
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUserData"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn",
          "user"
        ],
        "additionalProperties": true
      },
      "LogoutData": {
        "type": "object",
        "properties": {
          "loggedOut": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "loggedOut"
        ],
        "additionalProperties": true
      },
...
      "LoginDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LoginData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
      "CurrentUserDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CurrentUserData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
      "LogoutDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LogoutData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
...
    }
  },
...
}